rviz panel: Don't add object marker if the wrong tab is selected
The panel maintains an interactive marker by which the user can move
objects around in the planning scene.
Previously, this marker was created/updated whenever an object
was selected in the "Scene Objects" tab and geometry updates are received.

However, the marker should only be visible if the user actually
*wants* to manipulate the selected object, i.e. the correct tab is actually
active right now.

Because of the broken condition, the marker reappeared after
the user had selected an object (the marker appeared), selected a different tab
(the marker disappears), and the panel's planning scene monitor
received a geometry update afterwards. (This includes TRANSFORM updates that
are triggered by almost every interaction with move_group)