0 joint_times imply published static joints
So do not wait for them.

joints that are not published at all do not have an entry in joint_time_.
0 on the other hand implies that the joint was published on
tf_static and thus should be considered updated.

This resolves an edge-case I hit where the last joint in the list was a
floating joint (in this specific setup) published on tf_static.