further locking adaptations
As @rhaschke suggested, reading dt_state_update_ and
last_robot_state_update_wall_time_ does not lead to logic errors,
but at most to a skipped or redundant update on corrupted data.
Alternatively we could be on the safe side and turn both variables into
std::atomic, but would effectively require locks on every read.

Instead, only set state_update_pending_ as an atomic.
It's lockfree, so there is no reason not to when it's volatile already)
for now.